--- src/app/(admin)/admin/_components/Organization/SelectTherapies.js
@@ -1,20 +1,0 @@
-import { Loading, SelectArrayInput, required, useGetList } from 'react-admin';
-import { RESOURCES } from '@admin/_lib/consts';
-import PropTypes from 'prop-types';
-
-export function SelectTherapies({ isActive }) {
-  const { data: therapies, isLoading } = useGetList(RESOURCES.therapy);
-  if (isLoading) return <Loading />;
-  return (
-    <SelectArrayInput
-      label="Типи"
-      source="therapies"
-      validate={isActive && required()}
-      choices={therapies.map(therapy => ({ id: therapy.id, name: therapy.title }))}
-    />
-  );
-}
-
-SelectTherapies.propTypes = {
-  isActive: PropTypes.bool,
-};
