--- src/app/(admin)/admin/_components/Organization/index.js
@@ -1,3 +1,0 @@
-export * from './List';
-export * from './Show';
-export * from './Create';
