+++ src/app/(admin)/admin/_components/ServiceProvider/Organization/Create.js
@@ -1,6 +1,15 @@
 import { transformOrganizationData } from '@admin/_utils/transformOrganizationData';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Create, FormDataConsumer, NumberInput, SimpleForm, TextInput, required } from 'react-admin';
+import {
+  Create,
+  FormDataConsumer,
+  NumberInput,
+  SimpleForm,
+  TextInput,
+  required,
+  useNotify,
+  useRedirect,
+} from 'react-admin';
 import { FormFieldWrapper } from '../../FormFieldWrapper';
 import { ActivationForm } from '../ActivationForm';
 import { ContactsForm } from '../ContactsForm';
@@ -9,12 +18,25 @@ import { OrganizationTypesSelect } from '../OrganizationTypesSelect';
 import { ServicesForm } from '../ServicesForm';
 import { AddressesForm } from '../AddressesForm';
 import { organizationCreateValidationSchema } from '../../../_lib/validationSchemas/organizationSchema';
+import { RESOURCES, SUCCESS_NOTIFICATIONS } from '../../../_lib/consts';
 
 const fieldGroupClass = 'flex w-full flex-col md:flex-row md:gap-6 [&>*]:flex-grow';
 
 export function OrganizationCreate() {
+  const notify = useNotify();
+  const redirect = useRedirect();
+
+  const handleSuccess = () => {
+    notify(SUCCESS_NOTIFICATIONS.created);
+    redirect(`/${RESOURCES.organization}`);
+  };
+
+  const handleError = error => {
+    notify(error.message);
+  };
+
   return (
-    <Create transform={transformOrganizationData}>
+    <Create mutationOptions={{ onSuccess: handleSuccess, onError: handleError }} transform={transformOrganizationData}>
       <SimpleForm resolver={zodResolver(organizationCreateValidationSchema)}>
         <FormDataConsumer>
           {({ formData }) => {
@@ -22,14 +44,14 @@ export function OrganizationCreate() {
             const unnecessaryForDraft = formData.isActive && required();
             return (
               <>
-                <FormFieldWrapper title={'General info'}>
+                <FormFieldWrapper title={'Основна інформація'}>
                   <div className={fieldGroupClass}>
                     <TextInput source="name" label="Назва організації" validate={required()} />
                   </div>
                   <OrganizationTypesSelect fullWidth validate={required()} label={'Тип організації'} />
                 </FormFieldWrapper>
 
-                <FormFieldWrapper title={'Details'} className="mt-3">
+                <FormFieldWrapper title={'Деталі'} className="mt-3">
                   <div className={fieldGroupClass}>
                     <NumberInput
                       name={'yearsOnMarket'}
