+++ src/app/(admin)/admin/_components/ServiceProvider/Specialist/Edit.js
@@ -9,58 +9,11 @@ import { GeneralInfoEdit } from '../GeneralInfoEdit';
 import { FormFieldWrapper } from '../../FormFieldWrapper';
 import { GenderSelect } from '../GenderSelect';
 import { FormatOfWorkSelect } from '../FormatOfWorkSelect';
-
-const { FormatOfWork } = require('@prisma/client');
-
-const transformData = data => {
-  // console.log({ data: JSON.stringify(data) });
-  const therapiesToConnect = data.therapiesIds?.map(id => ({ id })) ?? [];
-  const specializationsToConnect = data.specializationsIds?.map(id => ({ id }));
-  const addressesToConnect = data.addresses?.filter(address => address.id).map(address => ({ id: address.id })) ?? [];
-  const addressesToCreate =
-    data.addresses
-      ?.filter(address => !address.id)
-      .map(address => ({
-        ...address,
-        district: { connect: { id: address.districtId } },
-        districtId: undefined,
-      })) ?? [];
-
-  const unselectedAddresses =
-    data.addressesIds?.filter(addressId => !addressesToConnect.some(address => address.id === addressId)) ?? [];
-  // if formatOfWork is ONLINE, we need to delete all connected addresses
-  const addressesToDelete =
-    data.formatOfWork !== FormatOfWork.ONLINE ? unselectedAddresses.map(id => ({ id })) ?? [] : {};
-
-  return {
-    ...data,
-    specializationsIds: undefined,
-    therapiesIds: undefined,
-    addressesIds: undefined,
-    therapies: {
-      set: [],
-      connect: therapiesToConnect,
-    },
-    specializations: {
-      set: [],
-      connect: specializationsToConnect,
-    },
-    addresses: {
-      connect: addressesToConnect,
-      create: addressesToCreate,
-      deleteMany: addressesToDelete,
-    },
-  };
-};
+import { transformSpecialistEditData } from '../../../_utils/transformSpecialistEditData';
 
 export function SpecialistEdit() {
   return (
-    <Edit
-      title={'Редагувати дані спеціаліста'}
-      // className="w-[800px]"
-      transform={transformData}
-      mutationMode="pessimistic"
-    >
+    <Edit title={'Редагувати дані спеціаліста'} transform={transformSpecialistEditData} mutationMode="pessimistic">
       <SimpleForm mode="all" reValidateMode="onChange" resolver={zodResolver(specialistEditValidationSchema)}>
         <FormDataConsumer>
           {({ formData }) => {
