--- src/app/(admin)/admin/_components/Specialist/Contacts.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import { FormFieldWrapper } from '@admin/components/FormFieldWrapper';
-import { SpecialistFormFields, SpecialistFormSections } from '@admin/_lib/specialistData';
-import { TextInputList } from '@admin/components/TextInputList';
-
-export function Contacts() {
-  const { phone, email, website, instagram, facebook, youtube, linkedin, tiktok, viber, telegram } =
-    SpecialistFormFields;
-  const contactsList = [phone, email, website];
-  const socialMediaList = [instagram, facebook, youtube, linkedin, tiktok, viber, telegram];
-
-  return (
-    <FormFieldWrapper title={SpecialistFormSections.contacts} className="mt-3">
-      <div className="flex w-full flex-col md:flex-row md:gap-6 [&>*]:flex-grow">
-        <TextInputList textInputList={contactsList} />
-      </div>
-      <div className="flex w-full flex-col flex-wrap md:flex-row md:gap-6 [&>*]:flex-grow">
-        <TextInputList textInputList={socialMediaList} className="w-1/3" />
-      </div>
-    </FormFieldWrapper>
-  );
-}
