--- src/app/(admin)/admin/_components/Specialist/index.js
@@ -1,3 +1,0 @@
-export * from './Create';
-export * from './List';
-export * from './Show';
