+++ src/app/(admin)/admin/_utils/transformSpecialistFormData.js
@@ -6,22 +6,6 @@ const transformAddresses = placesArray =>
     district: { connect: { id: place.district } },
   }));
 
-// export const transformData = data => {
-//   console.log({ transformationData: data });
-//   return {
-//     ...data,
-//     specializations: {
-//       connect: data.specializations?.length ? mapIdArrayToIdObjects(data.specializations) : undefined,
-//     },
-//     addresses: {
-//       create: data.addresses?.length ? transformAddresses(data.addresses) : undefined,
-//     },
-//     therapies: {
-//       connect: data.therapies?.length ? mapIdArrayToIdObjects(data.therapies) : undefined,
-//     },
-//   };
-// };
-
 export const transformData = data => ({
   ...data,
   specializations: {
