--- src/lib/validationSchemas/specialistSchema.js
@@ -1,149 +1,0 @@
-import { z } from 'zod';
-import { FormatOfWork, Gender } from '@prisma/client';
-import { PHONE_REGEX } from '@/lib/consts';
-
-// ------------------ COMMON SECTION ---------------------
-const MESSAGES = {
-  requiredField: `Обов'язкове поле`,
-  unacceptableValue: 'Недопустиме значення',
-};
-
-const zString = z
-  .string({
-    required_error: MESSAGES.requiredField,
-    invalid_type_error: MESSAGES.requiredField,
-  })
-  .trim();
-
-const zStringWithMax = zString.max(128, {
-  message: 'Поле не повинно перевищувати 128 символів',
-});
-
-const zStringArray = zString.array().min(1, {
-  message: MESSAGES.requiredField,
-});
-
-const zYearsOfExperience = z
-  .number({
-    required_error: MESSAGES.requiredField,
-    invalid_type_error: MESSAGES.unacceptableValue,
-  })
-  .nonnegative()
-  .nullish();
-
-const specialistCore = z.object({
-  isActive: z.boolean().optional(),
-  surname: zStringWithMax.nullish(),
-  gender: zString.refine(val => Object.values(Gender).includes(val), {
-    message: MESSAGES.unacceptableValue,
-  }),
-  yearsOfExperience: zYearsOfExperience,
-  description: zString.nullish(),
-  formatOfWork: zString.refine(val => Object.values(FormatOfWork).includes(val), {
-    message: MESSAGES.unacceptableValue,
-  }),
-  isFreeReception: z.boolean(),
-  phone: zString
-    .refine(val => PHONE_REGEX.test(val), {
-      message: 'Введіть номер телефона у форматі +380XXXXXXXXX',
-    })
-    .nullish(),
-  email: zString.email().nullish(),
-  website: zString.url().nullish(),
-  addressesIds: zString.array().nullish(),
-});
-
-const createValidationSchema = (schemaUnion, defaultProperties) =>
-  z.intersection(schemaUnion, defaultProperties).superRefine((schema, ctx) => {
-    const { formatOfWork, isActive, addresses } = schema;
-
-    if (isActive && formatOfWork !== FormatOfWork.ONLINE && !addresses.length) {
-      ctx.addIssue({
-        code: 'custom',
-        message: 'Необхідно вказати мінімум одне місце надання послуг',
-        path: ['addresses'],
-      });
-    }
-
-    if (formatOfWork === FormatOfWork.ONLINE) {
-      return {
-        ...schema,
-        addresses: [],
-      };
-    }
-
-    return schema;
-  });
-
-// ------------------ EDIT SECTION ---------------------
-
-const zEditAddressSchema = z.object({
-  id: z.string().nullish(),
-  fullAddress: zStringWithMax,
-  nameOfClinic: z.string().nullish(),
-  districtId: z.string(),
-  district: z
-    .object({
-      id: z.string(),
-      name: zString,
-    })
-    .nullish(),
-});
-
-const editDefaultProps = z.object({
-  lastName: zStringWithMax,
-  firstName: zStringWithMax,
-  specializationsIds: zStringArray,
-});
-
-const activeSpecialistEditSchema = specialistCore.extend({
-  addresses: zEditAddressSchema
-    .array()
-    .min(1, {
-      message: MESSAGES.requiredField,
-    })
-    .default([]),
-  therapiesIds: zStringArray,
-  isActive: z.literal(true),
-});
-
-const draftSpecialistEditSchema = specialistCore.partial().extend({
-  addresses: zEditAddressSchema.array(),
-  isActive: z.literal(false),
-});
-
-const specialistSchemaEditUnion = z.discriminatedUnion('isActive', [
-  activeSpecialistEditSchema,
-  draftSpecialistEditSchema,
-]);
-
-export const specialistEditValidationSchema = createValidationSchema(specialistSchemaEditUnion, editDefaultProps);
-// ------------------ CREATE SECTION ---------------------
-
-const zCreateAddressSchema = z.object({
-  fullAddress: z.string().nullish(),
-  district: z.string().nullish(),
-  nameOfClinic: z.string().nullish(),
-});
-
-const defaultProps = z.object({
-  lastName: zStringWithMax,
-  firstName: zStringWithMax,
-  specializations: zStringArray,
-});
-
-const restProps = specialistCore.extend({
-  therapies: zStringArray,
-  addresses: zCreateAddressSchema.array().default([]),
-});
-
-const activeSpecialistSchema = restProps.extend({
-  isActive: z.literal(true),
-});
-
-const draftSpecialistSchema = restProps.partial().extend({
-  isActive: z.literal(false),
-});
-
-export const specialistSchemaUnion = z.discriminatedUnion('isActive', [activeSpecialistSchema, draftSpecialistSchema]);
-export const specialistValidationSchema = createValidationSchema(specialistSchemaUnion, defaultProps);
